Fixed auto layout issues, created new view controller for sign up screen, implemented log in, added button to dismiss modal overlay, made modal overlay full screen
